feat(types): add declarations for versioned manifest files and their types
Contains information on how to increment the schema revision
